SQLAlchemyDaniLopez/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0